requirements.txt
fastapi
pydantic_settings
starlette
uvicorn
loguru
PyJWT
pymysql
sqlalchemy
python-dotenv
python-multipart
python-gitlab
pymongo[srv]

celery[SQLAlchemy,redis]
docker
eventlet
flower
sshtunnel
pysmb
pyautogui

pyyaml
websocket